pyproject.toml
[project]
name = "zabbix-mcp-server"
version = "2.0.0"
description = "A comprehensive MCP server for Zabbix integration"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Zabbix MCP Server Contributors"}
]
keywords = ["zabbix", "mcp", "monitoring", "api", "server"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
"fastmcp>=v2.12.4",
"zabbix_utils>=2.0.3",
"python-dotenv>=1.1.1"
]
requires-python = ">=3.10"

[project.optional-dependencies]
igraph = ["igraph~=0.11.6"]

[project.urls]
Homepage = "https://github.com/mpeirone/zabbix-mcp-server"
Repository = "https://github.com/mpeirone/zabbix-mcp-server"
Issues = "https://github.com/mpeirone/zabbix-mcp-server/issues"
Documentation = "https://github.com/mpeirone/zabbix-mcp-server#readme"

[project.scripts]
zabbix-mcp-readonly = "zabbix_mcp_server.main_readonly:main"
zabbix-mcp-writable = "zabbix_mcp_server.main_writable:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
